% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_faithful_histogram.R
\name{mod_faithful_histogram}
\alias{mod_faithful_histogram}
\alias{mod_faithful_histogram_ui}
\alias{mod_faithful_histogram_server}
\title{Old Faithful histogram Shiny module}
\usage{
mod_faithful_histogram_ui(id)

mod_faithful_histogram_server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
UI and server function for an example Shiny module.
}
